refactoring: Taking out class Client in a separate file
